mini-services/smolvlm2_server/requirements.txt
fastapi
uvicorn
pydantic
pillow
torch
transformers>=5.5.4
num2words
